refactor: hoist the completed-change merge rule onto CompletedChange
PartitionGroupConfiguration resolved an equal-version lastChange pair with a
private helper: same change id, earliest completedAt wins; different id, the
higher one wins. The rule is not specific to partition groups. It exists
because a graph change is completed without a coordinator -- every broker runs
completeGraphChangeIfDrained and mints its own record from its own view of when
the last operation finished -- so two brokers a moment apart disagree on
completedAt for the same change at the same version, and every other
equal-version field is resolved by keeping the receiver's copy, which would
leave that disagreement standing forever.

Any sub-configuration completing a change that way needs the same rule, so it
moves to CompletedChange where the next one can reach it. Pure move: the
resolution is unchanged, and PartitionGroupConfiguration's merge behaves
identically.
